+++ dev/ff_simulator.R
@@ -4,6 +4,7 @@
 options(ffscrapr.cache = "filesystem")
 
 pkgload::load_all()
+# library(ffsimulator)
 library(ffscrapr)
 library(dplyr)
 library(tidyr)