+++ inst/simulation_template_season.R
@@ -3,7 +3,8 @@
 #### PACKAGE LOAD ####
 options(ffscrapr.cache = "filesystem")
 
-library(ffsimulator)
+pkgload::load_all()
+# library(ffsimulator)
 library(ffscrapr)
 library(dplyr)
 library(tidyr)
@@ -13,11 +14,12 @@ library(ggplot2)
 library(cli)
 library(progressr)
 
+
 #### PARAMETERS ####
 
 conn <- mfl_connect(season = 2021, league_id = 22627)
 
-n_seasons <-  100 # any number of seasons
+n_seasons <-  500 # any number of seasons
 n_weeks <-  14 # any number of weeks per season
 best_ball <-  FALSE # or TRUE
 injury_model <-  "simple" # or none
@@ -27,25 +29,18 @@ seed <- 613
 set.seed(seed)
 
 #### IMPORT DATA ####
-cli::cli_rule("Starting simulation {Sys.time()}")
 
-cli::cli_process_start(msg = "Importing data")
+league_info <- ffscrapr::ff_league(conn = conn)
 
-league_info <- ff_league(conn = conn)
-
-scoring_history <- ff_scoringhistory(conn = conn, season = base_seasons)
+scoring_history <- ffscrapr::ff_scoringhistory(conn = conn, season = base_seasons)
 
 latest_rankings <- ffs_latest_rankings()
 
 rosters <- ffs_rosters(conn = conn)
 
-lineup_constraints <- ff_starter_positions(conn = conn)
-
-cli::cli_process_done(msg_done = "Importing data...done! {Sys.time()}")
+lineup_constraints <- ffscrapr::ff_starter_positions(conn = conn)
 
 #### GENERATE PROJECTIONS ####
-cli::cli_process_start(msg = "Generating Projections")
-
 adp_outcomes <- ffs_adp_outcomes(
   scoring_history = scoring_history,
   injury_model = injury_model
@@ -57,36 +52,19 @@ projected_scores <- ffs_generate_projections(
   n_weeks = n_weeks,
   rosters = rosters
 )
-
-cli::cli_process_done(msg_done = "Generating Projections...done! {Sys.time()}")
-
-
 #### CALCULATE ROSTER SCORES AND OPTIMAL LINEUPS ####
-cli::cli_process_start(msg = "Calculating Roster Scores")
-
 roster_scores <- ffs_score_rosters(
   projected_scores = projected_scores,
   rosters = rosters
 )
-
-cli::cli_process_done(msg_done = "Calculating Roster Scores...done! {Sys.time()}")
-
-cli::cli_process_start("Optimizing Lineups")
-
 optimal_scores <-
   ffs_optimise_lineups(
     roster_scores = roster_scores,
     lineup_constraints = lineup_constraints,
     best_ball = best_ball,
     pos_filter = c("QB","RB","WR","TE")
-    )
-
-cli::cli_process_done("Optimizing Lineups...! {Sys.time()}")
-
-
+  )
 #### SUMMARISE SIMULATION DATA ####
-cli::cli_process_start(msg = "Summarising Simulation Data")
-
 schedules <- ffs_build_schedules(
   n_teams = length(unique(rosters$franchise_id)),
   n_seasons = n_seasons,
@@ -97,10 +75,9 @@ summary_week <- ffs_summarise_week(optimal_scores, schedules)
 summary_season <- ffs_summarise_season(summary_week)
 summary_simulation <- ffs_summarise_simulation(summary_season)
 
-cli::cli_process_done(msg_done = "Summarising Simulation Data...done! {Sys.time()}")
-
 simulation <- structure(
   list(
+    summary_simulation = summary_simulation,
     summary_season = summary_season,
     summary_week = summary_week,
     league_info = league_info,
@@ -110,12 +87,13 @@ simulation <- structure(
       best_ball = best_ball,
       seed = seed,
       injury_model = injury_model,
-      base_seasons = base_seasons
+      base_seasons = base_seasons,
+      scrape_date = latest_rankings$scrape_date[[1]]
     )
   ),
   class = "ff_simulation"
 )
 
-autoplot(out, type = "wins")
-autoplot(out, type = "points")
-autoplot(out, type = "ranks")
+autoplot(simulation, type = "wins")
+autoplot(simulation, type = "points")
+autoplot(simulation, type = "rank")